+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project follows [Semantic Versioning](http://semver.org) guidelines.
 * `[rc.2]` — Fixes bug with server creation API endpoint that would fail to validate `allocation.default` correctly.
 * `[rc.2]` — Fix data integrity exception occuring due to invalid data being passed to server creation service on the API.
 * `[rc.2]` — Fix data integrity exception that could occur when an email containing non-username characters was passed.
+* `[rc.2]` — Fix data integrity exception occurring when no default value is provided for an egg variable.
 
 ### Added
 * Added ability to search the following API endpoints: list users, list servers, and list locations.